+++ Landscape/game_mod/img_loader.hpp
@@ -1,4 +1,4 @@
 #pragma once
 
 
-void ImgLoaderPatch(class PatcherInstance& p, HMODULE hModule);
+extern "C" __declspec(dllexport) void ImgLoaderPatch(class PatcherInstance& p, HMODULE hModule);